Category repository updated

+++ Flx.Domain/Models/SubCategory.cs
@@ -8,8 +8,8 @@ namespace Flx.Domain.Domains
 {
     public class SubCategory
     {
-        public int Id { get; set; }
-        public int CategoryId { get; set; }
+        public long Id { get; set; }
+        public long CategoryId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
     }
